requirements.txt
pymongo==4.5.0
python-dotenv==1.0.0
google-generativeai==0.3.0
PyGithub==2.1.1
requests==2.31.0
httpx==0.27.0
fastapi==0.110.0
uvicorn[standard]==0.28.0
asgiref==3.7.2
